fix potential type mismatch between format specifier and argument
Cast argument of type uint64_t to unsigned long long to match the %llu
format specifier on platforms where this might not be the case.

ok thomas

+++ lib/delta.c
@@ -323,7 +323,8 @@ got_delta_apply_in_mem(uint8_t *base_buf, size_t base_bufsz,
 	if (err == NULL && *outsize != result_size)
 		err = got_error_fmt(GOT_ERR_BAD_DELTA,
 		    "delta application result size mismatch: actual: %zd "
-		    "expected: %llu", *outsize, result_size);
+		    "expected: %llu", *outsize,
+		    (unsigned long long)result_size);
 
 	return err;
 }
@@ -392,7 +393,8 @@ got_delta_apply(FILE *base_file, const uint8_t *delta_buf,
 	if (err == NULL && *outsize != result_size)
 		err = got_error_fmt(GOT_ERR_BAD_DELTA,
 		    "delta application result size mismatch: actual: %zd "
-		    "expected: %llu", *outsize, result_size);
+		    "expected: %llu", *outsize,
+		    (unsigned long long)result_size);
 
 	if (memstream != NULL) {
 		if (fclose(memstream) == EOF && err == NULL)
